requirements.txt
huggingface_hub
flask
pyyaml